requirements.txt
numpy>=1.24.0
opencv-python-headless>=4.8.0
pyyaml>=6.0
# Optional for real model inference
torch>=2.0.0
PySide6>=6.6.0

